requirements.txt
aiogram
python-dotenv
openai
httpx
pdflatex
pydub
fastapi
uvicorn
